backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0

# Database
sqlmodel==0.0.22
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.10

# Redis
redis==5.2.0
hiredis==3.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.12
argon2-cffi==23.1.0

# Email
aiosmtplib==3.0.2
email-validator==2.2.0

# Rate Limiting
slowapi==0.1.9

# HTTP Client
httpx==0.27.2
aiohttp==3.10.10

# Validation & Serialization
pydantic==2.9.2
pydantic-settings==2.6.0

# Utilities
python-dotenv==1.0.1
pytz==2024.2

# Monitoring & Logging
python-json-logger==3.1.0

# WebSocket & Exchange APIs
websockets==12.0
ccxt==4.2.25
python-binance==1.0.19
bybit==0.3.1
krakenex==2.2.0
aioredis==2.0.1

# Optional: for advanced rate limiting
aiolimiter==1.1.0